Move Dimensions trait to geometry module
