+++ embedded-graphics/src/fonts/font6x12.rs
@@ -90,9 +90,8 @@ pub type Font6x12<'a, C> = FontBuilder<'a, C, Font6x12Conf>;
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::drawable::Dimensions;
     use crate::fonts::Font;
-    use crate::geometry::{Point, Size};
+    use crate::geometry::{Dimensions, Point, Size};
     use crate::mock_display::MockDisplay;
     use crate::pixelcolor::BinaryColor;
     use crate::style::WithStyle;
