+++ embedded-graphics/src/image/image_bmp.rs
@@ -1,6 +1,6 @@
 use super::ImageFile;
-use crate::drawable::{Dimensions, Drawable, Pixel};
-use crate::geometry::{Point, Size};
+use crate::drawable::{Drawable, Pixel};
+use crate::geometry::{Dimensions, Point, Size};
 use crate::pixelcolor::raw::{LittleEndian, RawData, RawDataIter};
 use crate::pixelcolor::PixelColor;
 use crate::transform::Transform;
