+++ embedded-graphics/src/primitives/circle.rs
@@ -1,8 +1,8 @@
 //! The circle primitive
 
-use super::super::drawable::*;
-use super::super::transform::*;
-use crate::geometry::{Point, Size};
+use super::super::drawable::{Drawable, Pixel};
+use super::super::transform::Transform;
+use crate::geometry::{Dimensions, Point, Size};
 use crate::pixelcolor::PixelColor;
 use crate::primitives::Primitive;
 use crate::style::Style;
@@ -270,7 +270,6 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::drawable::Dimensions;
     use crate::pixelcolor::BinaryColor;
 
     #[test]
