+++ embedded-graphics/src/primitives/line.rs
@@ -1,8 +1,8 @@
 //! The line primitive
 
-use super::super::drawable::*;
-use super::super::transform::*;
-use crate::geometry::{Point, Size};
+use super::super::drawable::{Drawable, Pixel};
+use super::super::transform::Transform;
+use crate::geometry::{Dimensions, Point, Size};
 use crate::pixelcolor::PixelColor;
 use crate::primitives::Primitive;
 use crate::style::Style;
